+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## 1.10.0.0
 
+* Add `psDRepDistrG`.
+* Rename `ensPParams` to `ensCurPParams`.
+* Add `ToJSON` instance for `RatifyState`
+* Change `ToJSON` instance for `ConwayGovState`:
+  * Add `"nextRatifyState"` field
+  * Rename `"ratify"` to `"enactState"`
+  * Rename `"gov"` to `"proposals"`
+* Fix `ToJSON` instance for `EnactState`:
+  * Current PParams were wrongfully used for `"prevPParams"`.
+  * Remove `"treasury"` and `"withdrawals"` as those are temporary bindings needed only
+    for `ENACT` rule
 * Add an anchor argument to `ResignCommitteeColdTxCert`
 * Prevent invalid previous gov-action ids in proposals #3768
   * Also, add lenses